cargo clippy

+++ packets/src/reader.rs
@@ -11,7 +11,7 @@ impl<R: AsyncRead + Unpin + Send + Sync> AsyncReadPacket for R {
         }
 
         // extract the encoded packets id and validate if it is expected
-        let packet_id = self.read_varint().await? as usize;
+        let packet_id = self.read_varint().await?;
         let expected_packet_id = T::get_packet_id();
         if packet_id != expected_packet_id {
             return Err(Error::IllegalPacketId {
